// src/contracts/generated/ViewModel/patreonSettingsViewModel.ts
/* Auto Generated */

import { PatreonBannerDisplayType } from './../Enum/patreonBannerDisplayType';

export interface PatreonSettingsViewModel {
  displayType: PatreonBannerDisplayType;
  foregroundColour: string;
  backgroundColour: string;
  backgroundOpacity: number;
  marqueSpeed: number;
  verticalListSpeed: number;
  oneAtATimeSpeed: number;
  isProfilePicRounded: boolean;
  profilePicRoundedValue: number;
  panelForegroundColour: string;
  panelBackgroundColour: string;
  panelBackgroundOpacity: number;
  panelVerticalListSpeed: number;
  panelUseDefaultBackground: boolean;
  panelCustomBackgroundImageUrl: string;
  panelIsProfilePicRounded: boolean;
  panelProfilePicRoundedValue: number;
  lastModifiedDate: Date;
}